+++ SmallBeachFarm/Patches/FarmPatcher.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using HarmonyLib;
 using Microsoft.Xna.Framework;
@@ -6,6 +7,7 @@ using Pathoschild.Stardew.Common.Patching;
 using Pathoschild.Stardew.SmallBeachFarm.Framework;
 using StardewModdingAPI;
 using StardewValley;
+using StardewValley.Locations;
 using SObject = StardewValley.Object;
 
 namespace Pathoschild.Stardew.SmallBeachFarm.Patches
@@ -33,6 +35,16 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Patches
 
         /// <summary>Whether the mod is currently applying patch changes (to avoid infinite recursion),</summary>
         private static bool IsInPatch;
+
+        /// <summary>The arrival tiles by farm type when the player warps from the <see cref="IslandWest"/> obelisk, with a <c>-1</c> key for the default.</summary>
+        /// <remarks>Derived from <see cref="IslandWest.performAction"/>.</remarks>
+        [Obsolete("This will no longer be needed in Stardew Valley 1.5.5.")]
+        private static readonly Dictionary<int, Point> IslandWarpTargets = new()
+        {
+            [Farm.fourCorners_layout] = new Point(48, 39),
+            [Farm.beach_layout] = new Point(81, 29),
+            [-1] = new Point(48, 7)
+        };
 
 
         /*********
@@ -134,6 +146,14 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Patches
             // change background track
             if (FarmPatcher.ShouldUseBeachMusic())
                 Game1.changeMusicTrack("ocean", music_context: Game1.MusicContext.SubLocation);
+
+            // fix island totem warp
+            // TODO: remove in Stardew Valley 1.5.5
+            if (FarmPatcher.IsOnIslandWarpArrivalTile())
+            {
+                var tile = __instance.GetMapPropertyPosition("WarpTotemEntry", 48, 7);
+                Game1.player.Position = new Vector2(tile.X * Game1.tileSize, tile.Y * Game1.tileSize);
+            }
         }
 
         /// <summary>A method called via Harmony after <see cref="Farm.resetSharedState"/>.</summary>
@@ -178,6 +198,18 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Patches
         private static bool ShouldUseBeachMusic()
         {
             return FarmPatcher.UseBeachMusic && !Game1.isRaining;
+        }
+
+        /// <summary>Whether the player is on the arrival tile for the <see cref="IslandWest"/> obelisk warp.</summary>
+        [Obsolete("This will no longer be needed in Stardew Valley 1.5.5.")]
+        private static bool IsOnIslandWarpArrivalTile()
+        {
+            Point curTile = Utility.Vector2ToPoint(Game1.player.Position / Game1.tileSize);
+
+            if (!FarmPatcher.IslandWarpTargets.TryGetValue(Game1.whichFarm, out Point arrivalTile))
+                arrivalTile = FarmPatcher.IslandWarpTargets[-1];
+
+            return curTile == arrivalTile;
         }
     }
 }