+++ SkipIntro/docs/release-notes.md
@@ -2,7 +2,7 @@
 
 # Release notes
 ## Upcoming release
-* Updated for Stardew Valley 1.5.5.
+* Internal changes for compatibility with the upcoming Stardew Valley 1.5.5.
 
 ## 1.9.1
 Released 27 March 2021 for SMAPI 3.9.5 or later.